backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langgraph==0.2.45
langchain==0.3.15
langchain-openai==0.2.8
python-dotenv==1.0.1
psutil==5.9.8